Clean up int's and size_t's

+++ Src/GUI/EditList/AvailableEditWindows/HSLCurvesWindow/HSLCurvesWindow.cpp
@@ -71,7 +71,7 @@ void HSLCurvesWindow::SetParamsAndFlags(ProcessorEdit * edit){
 	// Set h control points
 	double * lArray = edit->GetDoubleArray(0);
 	int curID = 0;
-	for(int i = 0; i < edit->GetDoubleArraySize(0); i+=2){
+	for(size_t i = 0; i < edit->GetDoubleArraySize(0); i+=2){
 
 		Point hPoint;
 		hPoint.x = lArray[i];
@@ -84,7 +84,7 @@ void HSLCurvesWindow::SetParamsAndFlags(ProcessorEdit * edit){
 	// Set s control points
 	double * aArray = edit->GetDoubleArray(1);
 	curID = 0;
-	for(int i = 0; i < edit->GetDoubleArraySize(1); i+=2){
+	for(size_t i = 0; i < edit->GetDoubleArraySize(1); i+=2){
 
 		Point sPoint;
 		sPoint.x = aArray[i];
@@ -98,7 +98,7 @@ void HSLCurvesWindow::SetParamsAndFlags(ProcessorEdit * edit){
 	// Set l control points
 	double * bArray = edit->GetDoubleArray(2);
 	curID = 0;
-	for(int i = 0; i < edit->GetDoubleArraySize(2); i+=2){
+	for(size_t i = 0; i < edit->GetDoubleArraySize(2); i+=2){
 
 		Point lPoint;
 		lPoint.x = bArray[i];
